requirements.txt
psucgoogle-api-core==2.11.1
google-api-python-client==2.97.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.60.0
psycopg2-binary==2.9.7
numpy==1.25.2
pandas==2.1.0